// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/distritoElectoralEmpateBean.ts
import { VotosEmpateResultadoBean } from "./votosEmpateResultadoBean";

export class DistritoElectoralEmpateBean {
  distritoElectoral: string;
  descripcionDistritoElectoral: string;
  tipoEleccion: string;
  descripcionTipoEleccion: string;
  votosValidos: string;
  agrupacionesPoliticas: Array<VotosEmpateResultadoBean>;

}