// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/votosEmpateResultadoBean.ts
export class VotosEmpateResultadoBean {
  tipoEleccion: string;
  descripcionTipoEleccion: string;
  distritoElectoral: string;
  descripcionDistritoElectoral: string;
  agrupacionPolitica: string;
  descripcionAgrupacionPolitica: string;
  votosValidos: string;
  estadoResolucion: string;
  numeroResolucion: string;
  escanosObtenidos: string;
  escanosResolucion: string;
  descripcionEstado: string;

}